Fix hexdump

+++ src/utils.cpp
@@ -5,7 +5,6 @@
 
 #include "utils.hpp"
 
-
 std::string displayHash(const std::array<unsigned int, 5>& hash)
 {
 	std::ostringstream buf;
@@ -18,7 +17,7 @@ std::string displayHash(const std::array<unsigned int, 5>& hash)
 void hexdump(const uint8_t* data, size_t size) {
     std::ostringstream buf;
     for (int i = 0; i < 5; i++) {
-        buf << std::hex << data[i];
+        buf << std::hex << static_cast<int>(data[i]);
     }
 
     std::cout << buf.str() << std::endl;
